+++ src/global_gen.rs
@@ -1,10 +1,11 @@
-//! Default generator and entry point functions
+//! Default generator and entry point functions.
 
 #![cfg(feature = "global_gen")]
+#![cfg_attr(docsrs, doc(cfg(feature = "global_gen")))]
 
 use std::sync;
 
-use crate::{Uuid, V7Generator};
+use crate::Uuid;
 use inner::GlobalGenInner;
 
 /// Returns the lock handle of process-wide global generator, creating one if none exists.
@@ -30,7 +31,6 @@ fn lock_global_gen() -> sync::MutexGuard<'static, GlobalGenInner> {
 ///
 /// let uuid_string: String = uuid7::uuid7().to_string();
 /// ```
-#[cfg_attr(docsrs, doc(cfg(feature = "global_gen")))]
 pub fn uuid7() -> Uuid {
     lock_global_gen().get_mut().generate()
 }
@@ -43,7 +43,6 @@ pub fn uuid7() -> Uuid {
 /// let uuid = uuid7::uuid4();
 /// println!("{}", uuid); // e.g., "2ca4b2ce-6c13-40d4-bccf-37d222820f6f"
 /// ```
-#[cfg_attr(docsrs, doc(cfg(feature = "global_gen")))]
 pub fn uuid4() -> Uuid {
     lock_global_gen().get_mut().generate_v4()
 }
@@ -52,7 +51,7 @@ mod inner {
     use rand::rngs::{adapter::ReseedingRng, OsRng};
     use rand_chacha::ChaCha12Core;
 
-    use super::V7Generator;
+    use crate::generator::{with_rand08, V7Generator};
 
     /// The type alias for the random number generator of the global generator.
     ///
@@ -60,7 +59,7 @@ mod inner {
     /// emulate the strategy used by [`rand::rngs::ThreadRng`].
     ///
     /// [`rand::rngs::ThreadRng`]: https://docs.rs/rand/0.8.5/rand/rngs/struct.ThreadRng.html
-    type GlobalGenRng = ReseedingRng<ChaCha12Core, OsRng>;
+    type GlobalGenRng = with_rand08::Adapter<ReseedingRng<ChaCha12Core, OsRng>>;
 
     /// A thin wrapper to reset the state when the process ID changes (i.e., upon Unix forks).
     #[derive(Debug)]
@@ -78,7 +77,7 @@ mod inner {
             Self {
                 #[cfg(unix)]
                 pid: std::process::id(),
-                generator: V7Generator::new(ReseedingRng::new(rng, 1024 * 64, OsRng)),
+                generator: V7Generator::with_rand08(ReseedingRng::new(rng, 1024 * 64, OsRng)),
             }
         }
     }